refactor classes into corresponding packages
