+++ src/test/java/seedu/badmaths/storage/NotesFileParserTest.java
@@ -1,4 +1,4 @@
-package seedu.badmaths;
+package seedu.badmaths.storage;
 
 import org.junit.jupiter.api.Test;
 import seedu.badmaths.note.Note;
